Adjust the WikiUpdater API
Rather than creating a ComponentIndex internally, the WikiUpdater
now expects one to be injected as an argument to the update method.

As such, the Info main method now creates the ComponentIndex first,
then the WikiUpdater, and finally passes the former to the latter.

The WikiUpdater now uses the g:a:v of the component for the summary
message, rather than the g:a:v of the project as previously.

These changes will pave the way for the WikiUpdater to support
upload of multiple projects at once, in a non-destructive way.